+++ src/js/components/notification.js
@@ -1,6 +1,9 @@
+import Container from '../mixin/container';
 import {$, append, apply, closest, css, parent, pointerEnter, pointerLeave, remove, startsWith, toFloat, Transition, trigger} from 'uikit-util';
 
 export default {
+
+    mixins: [Container],
 
     functional: true,
 
@@ -33,8 +36,8 @@ export default {
 
     created() {
 
-        const container = $(`.${this.clsContainer}-${this.pos}`, this.$container)
-            || append(this.$container, `<div class="${this.clsContainer} ${this.clsContainer}-${this.pos}" style="display: block"></div>`);
+        const container = $(`.${this.clsContainer}-${this.pos}`, this.container)
+            || append(this.container, `<div class="${this.clsContainer} ${this.clsContainer}-${this.pos}" style="display: block"></div>`);
 
         this.$mount(append(container,
             `<div class="${this.clsMsg}${this.status ? ` ${this.clsMsg}-${this.status}` : ''}">